fix(network-monitor): avoid warning
tcp_set_state was getting called after its process was already dead,
generating a warning because we couldn't find its entry in map_interest.

We now log the calling function name when logging that error (so it will
be easier to find the culprit if we get other errors). Doing so required
to switch from a function call to a macro, in order to use `__func__`