new(file-system-monitor): add symbolic and hard links support
Hook into LSM `path_link` and `path_symlink` to get events about new
symbolic and hard links. As usual, we use kprobes as a fallback.

Refactor `get_path_str` to accept the `struct vfsmount *` as a second
parameter. This allows to use a single version of the function, which
can be used both for `struct dentry` and `struct path`.

Fix #64