fix(network-monitor): fix buffer index warnings
In network-monitor, the send and receive event have a BufferIndex inside
the data field, which contains the actual message. This is only used for
UDP messages, since we use it to intercept DNS.

The change to dynamic message sizes using buffer and indexes introduced
a regression in the TCP path: we weren't setting the data.len to 0,
resulting in a garbage len being sent. The buffer was empty though, so
we got a lot of warnings. This explains why the errors didn't start
right away when starting pulsar (first, we needed to send or receive the
first UDP message) and kept coming up afterwards (the len was set in the
UDP probe and never reset until the next UDP message)

Fix #139