fix(process-monitor): fix failing test on embedded systems
On busybox-based embedded systems, `which echo` returns `/bin/echo`,
which is a symlink to `/bin/busybox`.
On relative path execution, our `sched_process_exec` handler takes the
path from `bprm->filename`, which leads to the linked at file.
For this reason we need to use `canonicalize()` in the relative exec
test case.