new(ebpf): use longer iterations when bpf_loop available
The LOOP macro now takes an extra `max_unroll` parameter, which defines
the maximum loop unroll when bpf_loop is missing.
This allows in `get_path_str` to have a MAX_PATH_COMPONENTS of 100, but
unroll the loop only up to 20, since a bigger value would be rejected by
the verifier.