chore: Link with clang/lld, update to LLVM 17
Latest CI failures are connected to the fact that Rust 1.73 is
using new toolchains, which results in ld linker errors.

However, in this PR, we fix the linker errors by switching to lld,
because apart from fixing the original issue, it also speeds up
the build.

Also, since the new Rust is using LLVM 17, use LLVM 17 for building
our C probes as well.