feat: Make it clear which modules are enabled by default
Before this change, empty config of Pulsar was resulting in all
modules being enabled. It's quite inconvenient in case of smtp-notifier,
which requires additional configuration in order to work without
errors, and desktop-notifier, which isn't meant for servers. Both
of these modules are not going to be used by most of people.

This change adds the `enabled_by_default` flag to `ModuleDetails`
which has to be defined by every module. For all modules except
smtp-notifier, it's set to `true`.