chore: Allow to build with rustls
ring 0.17 supports RISC-V, so building Pulsar with rustls instead
of OpenSSL is possible.

Choice of the TLS library is exposed by the following features
in the `smtp-notifier` crate:

* `native-tls` - uses the native TLS library in the system, usually
  OpenSSL. Enabled by default.
* `boring-tls` - uses BoringSSL. Disabled by default.
* `rustls-tls` - uses rustls. Disabled by default.

When building Pulsar with Cargo, the native TLS is going to be used.

However, for binary builds for Github releases, rustls is being
used, since it's always statically built and doesn't require any
runtime dependencies from the users.