fix: Include the data length and src/dest info for all supported protocols
Before this change, we were missing data length in some network events,
as we were missing the assignment in some cgroup_skb program branches.

We were also inconsistent in the way we were retrieving the source and
destination information. Sometimes we were getting it from the `args`
map. Sometimes, directly from SKB. The latter is better, since it just
used the packet data available locally, inside the probe.

Also, take this fix as an opportunity to refactor to move the whole
SKB parsing logic into one function, to minimize the risk of such
mismatch in the future.

The other minor changes done through the refactor are:

* Removing the necessity of looking at `skb->sock` and having `bind`
  and `connect` maps.
  * The lookup for the process is done with `bpf_get_current_task_btf`.
    The whole logic of mapping a socket to a process therefore becomes
    obsolete.
  * Addresses and ports are retrieved directly from the packet headers
    (IP and UDP/TCP).