fix(network-filter): DNS event generation
First the amount of data read is now corresponding to the data len - 1
in the packet (excl IP and UDP headers). Review is welcome on this point
since that does not work if "len = output->data_len" (i.e. the exact size of
the data in the packet) but with "len = output->data_len - 1".

For the data reception, destination address retrieval of the incoming
packet is skipped. Actually we do not really care about this adress since
it should be an address of one of the network interface...

Co-authored-by: Pierre Ansel <pierre.ansel@seedfence.fr>