feat(xtask): Unify `test` and `cross` subcommands
Having `cross` as a subcommand is somehow redundant - there is no real
benefit over using `cross` as a command directly if the caller is only
interested in builds.

However, it's beneficial to extend the `test` command with ability to
perform a cross build and use architest when the target doesn't match
the host.