feat: Include rootfs in the `Exec` event
`image` field of the `Exec` event is relative to the root filesystem
of the process. That makes it hard to determine the location of the
binary if the process is containerized (or, in general, runs inside a
mount namespace).

That's where the new `rootfs` comes with help. It contains the full path
to the root filesystem, which in case of containers has format like:

```
/root/var/lib/docker/btrfs/subvolumes/0991d46b6f686f22f06bddb6948073[...]
```

Using that information, we can still inspect that container layer even
when the process (or even container) are not running anymore.