fix: Handle offsets and their bounds in `buffer_append_str`
Before this change, we were trying to handle offsets by doing pointer
arithmetics (like `str + offset` where `str` is `char*` and `offset`
an integer).

Verifier in kernels <6.0 was not happy with that and was complaining
about unchecked bounds.

Fix that by making `buffer_append_str` aware of offsets and doing all
the bound checks there.